+++ test/examples/flappy_bird/browser-entry/playback/playback.loop.service.ts
@@ -1,0 +1,14 @@
+import { PlaybackAnimationFrameUnavailableError } from './playback.errors';
+
+/**
+ * Yields until the next browser animation frame.
+ *
+ * @returns Promise resolved on next animation frame.
+ */
+export function nextAnimationFrame(): Promise<void> {
+  if (typeof requestAnimationFrame !== 'function') {
+    throw new PlaybackAnimationFrameUnavailableError();
+  }
+
+  return new Promise((resolve) => requestAnimationFrame(() => resolve()));
+}
