+++ test/examples/flappy_bird/browser-entry/visualization/visualization.constants.ts
@@ -1,0 +1,19 @@
+/**
+ * Pixel side length for dotted negative-connection square markers.
+ */
+export const FLAPPY_NETWORK_DOTTED_CONNECTION_SQUARE_SIDE_PX = 2;
+
+/**
+ * Spacing multiplier used to tighten square-dot trail cadence.
+ */
+export const FLAPPY_NETWORK_DOTTED_CONNECTION_STEP_COMPACT_RATIO = 0.8;
+
+/**
+ * Line-width multiplier used to scale dotted-step spacing.
+ */
+export const FLAPPY_NETWORK_DOTTED_CONNECTION_WIDTH_SPACING_RATIO = 2.6;
+
+/**
+ * Tiny alignment epsilon to stabilize axis-aligned dot centering.
+ */
+export const FLAPPY_NETWORK_DOTTED_CONNECTION_ALIGNMENT_EPSILON = 0.01;
